Rename serial submodule to ports
